Feat: 엔티티 패키지 구조 개선
