+++ src/main/java/hackathon/soa/domain/participation/ParticipationController.java
@@ -2,8 +2,7 @@ package hackathon.soa.domain.participation;
 
 import hackathon.soa.common.JwtUser;
 import hackathon.soa.common.apiPayload.ApiResponse;
-import hackathon.soa.domain.search.dto.SearchResponseDTO;
-import hackathon.soa.entity.SegmentParticipationStatus;
+import hackathon.soa.entity.enums.SegmentParticipationStatus;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
