+++ src/main/java/hackathon/soa/domain/segment/SegmentService.java
@@ -14,6 +14,7 @@ import hackathon.soa.domain.segment.repository.CourseSegmentRepository;
 import hackathon.soa.domain.segment.repository.MoveSegmentRepository;
 import hackathon.soa.domain.segment.repository.StaySegmentRepository;
 import hackathon.soa.entity.*;
+import hackathon.soa.entity.enums.SegmentParticipationStatus;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
