+++ src/main/java/hackathon/soa/entity/CourseSegment.java
@@ -1,5 +1,6 @@
 package hackathon.soa.entity;
 
+import hackathon.soa.entity.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,8 +8,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "course_segment")
