--- src/main/java/hackathon/soa/entity/CourseStatus.java
@@ -1,5 +1,0 @@
-package hackathon.soa.entity;
-
-public enum CourseStatus {
-    IN_PROGRESS, COMPLETED
-}
