--- src/main/java/hackathon/soa/entity/Gender.java
@@ -1,5 +1,0 @@
-package hackathon.soa.entity;
-
-public enum Gender {
-    MALE, FEMALE, ALL
-}
