--- src/main/java/hackathon/soa/entity/MovementType.java
@@ -1,5 +1,0 @@
-package hackathon.soa.entity;
-
-public enum MovementType {
-    WALK, CAR, PUBLIC_TRANSIT, BUS, SUBWAY
-}
