+++ src/main/java/hackathon/soa/entity/SegmentParticipation.java
@@ -1,5 +1,7 @@
 package hackathon.soa.entity;
 
+import hackathon.soa.entity.common.BaseEntity;
+import hackathon.soa.entity.enums.SegmentParticipationStatus;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
