--- src/main/java/hackathon/soa/entity/SegmentParticipationStatus.java
@@ -1,5 +1,0 @@
-package hackathon.soa.entity;
-
-public enum SegmentParticipationStatus {
-    APPROVED, REJECTED, PENDING
-}
