+++ src/main/java/hackathon/soa/entity/common/BaseEntity.java
@@ -1,0 +1,22 @@
+package hackathon.soa.entity.common;
+
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+public abstract class BaseEntity {
+
+    @CreatedDate
+    private LocalDateTime createdAt; // 생성 일자
+
+    @LastModifiedDate
+    private LocalDateTime updatedAt; // 수정 일자
+}
