+++ src/main/java/hackathon/soa/entity/enums/CourseStatus.java
@@ -1,0 +1,5 @@
+package hackathon.soa.entity.enums;
+
+public enum CourseStatus {
+    IN_PROGRESS, COMPLETED
+}
