+++ src/main/java/hackathon/soa/entity/enums/Gender.java
@@ -1,0 +1,5 @@
+package hackathon.soa.entity.enums;
+
+public enum Gender {
+    MALE, FEMALE, ALL
+}
