+++ src/main/java/hackathon/soa/entity/enums/MovementType.java
@@ -1,0 +1,5 @@
+package hackathon.soa.entity.enums;
+
+public enum MovementType {
+    WALK, CAR, PUBLIC_TRANSIT, BUS, SUBWAY
+}
