+++ src/main/java/hackathon/soa/entity/enums/SegmentParticipationStatus.java
@@ -1,0 +1,5 @@
+package hackathon.soa.entity.enums;
+
+public enum SegmentParticipationStatus {
+    APPROVED, REJECTED, PENDING
+}
